pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hoancau-ai"
version = "0.1.0"
description = "HoanCau AI Resume Processor"
readme = "readme.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dependencies = [
    "python-dotenv>=0.21.0",
    "pdfminer.six==20231228",
    "PyPDF2>=3.0.0",
    "pymupdf>=1.22.0",
    "python-docx>=0.8.11",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "google-generativeai>=0.2.0",
    "requests>=2.31.0",
    "streamlit>=1.22.0",
    "fastapi>=0.95.0",
    "uvicorn[standard]>=0.22.0",
    "python-multipart>=0.0.5",
    "click",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.0.0"
]

[project.scripts]
cli-agent = "scripts.cli_agent:cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
